Pre/Post-processing client
Added a Python script that handles the pre/post-processing of image data. This file converts .tif images to IEEE 754 based .txt files, acts as a client on the host PC, sends data to the FPGA-based echo server, then receives the data back and converts it to scientific notation, and finally stores the results in a .txt file.