MUX IP
IP that reads both the reference image BRAM and the deformed image BRAM and alternates the images when new frames are received. The deformed image becomes the reference image and a new deformed image overwrites the original reference image that is no longer needed. This IP does not resemble a true MUX.