Build 1
First major build of the DICE accelerator. Handles < 1,000,000 frames of size 448x232 and up to 14 subsets (with square and circle subset shapes). This design is completed with an echo server setup through the Vivado SDK (which is in its own folder in the repo) along with the Python "processing" script that will interact with the echo server. The images need to be in the same file path as the python script in order to transmit. The computers IP address need to be changed to a static address like "192.168.1.12".